Changes on other main files

+++ src/PM.cpp
@@ -25,7 +25,7 @@ void PM(Mat_m &A, Mat_m &B, double &sigma, Vec_v &u, Vec_v &v) {
     double epsilon = 1.e-10;
     double delta = 0.05;
     double lambda = 0.1;
-    int s = ceil(log(4 * log(2 * A.cols() / delta) / (epsilon * delta)) / (2 * lambda));
+    int s = static_cast<int>(ceil(log(4 * log(2 * A.cols() / delta) / (epsilon * delta)) / (2 * lambda)));
 
     // Divide rows of A among processors
     int rows_per_proc = B.rows() / num_procs;
